Remove the actions altogether
Returning an array of actions instead of just calling appropriate
methods obfuscates the application logic. It also requires you to define
a separate action for every new thing you might want to do, which
results in a lot of boilerplate code.
The log messages generated from string representations of actions are
too generic to be satisfactory, e.g. the message might tell you that
"Antelope moves from a to b" but it doesn't describe that it did so to
escape from a Wolf. This problem could be solved by creating more
specialised actions with different string representations, but that
would result in even more boilerplate. Generating log messages inline is
more flexible and concise solution.